Add : 인증 기능 구현
* 단순 비밀번호 검사 기능만 추가했습니다.
