+++ MARKDOWN-BLOG/views/articles/index.ejs
@@ -14,7 +14,7 @@
     <div class="container">
         <h1 class="mb-4 mt-4">Glog</h1>
         <a href="/articles/new" class="btn btn-success">New Article</a>
-        <a href="/login" class="btn btn-primary">Login</a>
+        <a href="/auth/login" class="btn btn-primary">Login</a>
         <% articles.forEach(article=> { %>
             <div class="card mt-4">
                 <div class="card-body">
@@ -27,8 +27,7 @@
                     <div class="card-text mb-2">
                         <%= article.description %>
                     </div>
-                    <a href="articles/<%= article.slug %>" class="btn btn-primary">Read More</a>
-                    <a href="articles/edit/<%= article.id %>" class="btn btn-info">Edit</a>
+                    <a href="articles/<%= article.slug %>" class="btn btn-info">Read More</a>
                 </div>
             </div>
             <% }) %>
