+++ MARKDOWN-BLOG/views/logins/fail.ejs
@@ -9,7 +9,7 @@
 </head>
 <body>
     <h1>login Failed!</h1>
-    <a href="/login" class = "home-btn">Try again</a>
+    <a href="/auth/login" class = "home-btn">Try again</a>
     <a href="/">Home</a>
 </body>
 </html>
